test: split compat fixtures into focused bundles (#80)
