+++ compat-fixtures/src/main/scala-2.13/com/github/mercurievv/scalasemantic/compat/CallGraph.scala
@@ -1,0 +1,8 @@
+package com.github.mercurievv.scalasemantic.compat
+
+/** Flat call chain a -> b -> c for call-graph path-finding tests. */
+object Calls {
+  def a(): Int = b()
+  def b(): Int = c()
+  def c(): Int = 1
+}
