+++ compat-fixtures/src/main/scala-2.13/com/github/mercurievv/scalasemantic/compat/Overloads.scala
@@ -1,0 +1,7 @@
+package com.github.mercurievv.scalasemantic.compat
+
+/** Overloaded methods sharing a name and owner — exercises findOverloads. */
+object Overloads {
+  def over(x: Int): Int    = x
+  def over(x: String): String = x
+}
