+++ compat-fixtures/src/main/scala-3/com/github/mercurievv/scalasemantic/compat/Generics.scala
@@ -1,0 +1,5 @@
+package com.github.mercurievv.scalasemantic.compat
+
+/** Generic typeclass definition — used by the Implicits bundle and resolve-implicits tests. */
+trait Show[A]:
+  def show(a: A): String
